refactor: centralize embed configs
