+++ src/commands/ping/command.js
@@ -2,7 +2,7 @@
 ### Zweck: Slash-Command /ping – zeigt Ping, Roundtrip, Uptime und Serverzeit als Embed.
 */
 import { EmbedBuilder } from 'discord.js';
-import { FOOTER } from '../../util/footer.js';
+import { FOOTER } from '../../util/embeds/footer.js';
 
 export default {
   name: 'ping',
