+++ src/util/embeds/color.js
@@ -1,0 +1,4 @@
+/*
+### Zweck: Zentrale Embed-Farbe (Twitch-Lila).
+*/
+export const COLOR = 0x6441A5;
