+++ src/util/embeds/footer.js
